odop: Add a new class for tracking runs as benchmark records
ODOP keeps track of individual runs. The idea is that after the end of a
run the final results can be saved aside, compressed and also annotated
if needed for benchmarking or archival purposes.

A benchmark record should hold the following information:

- the name of the ODOP run
- the used ODOP con fig
- metrics captured during the run using OdopObs
- total execution time
- runtime environment

This adds only the basic class with initial loading logic. The
conversion of ODOP runs into benchmark records will be added in a
follow-up commit.