odop/benchmark: Add conversion of an ODOP run into a benchmark record
During a run ODOP collects a number of metrics used not just for
scheduling. They are vital when wanting to benchmark the effect of ODOP
on an application.

Currently OdobObs uses Tinyflux[0] library for storing time-series
metrics. These are stored per-node in a multi-node set-up. For more
convenient post-processing, it seems better to have all the metrics from
all nodes in the same time series. This is merge is done using a heap
queue.

In case the amount of metrics is already quite high, the recommendation
for tinyflux is to split a database into multiple files to keep the
query speeds reasonable[1].

[0] https://github.com/citrusvanilla/tinyflux
[1] https://tinyflux.readthedocs.io/en/latest/tips.html#dealing-with-growing-datasets