benchmark/MHD: Add generator for scaling experiments
The generator takes a number of parameters and based on these creates a
series of experiments for the weak and strong scaling benchmarks. The
files for the experiments are taken from a provided pencil-code
repository.

Specific characteristics of the generated experiments:

- the domain size aims to maximize utilization of GPU memory,
- the size of the domain is uniform,
- the domain is distributed across GPUs only across the Y-axis.

The distribution over Y-axis is an arbitrary one and I recommend
exploring other types of distributions.

The formulas for calculating the problem size (i.e., memory cost) of a
grid and finding a maximum grid for an amount of memory are exposed
through own commands.